NMS-9120: Persist string attributes over the TCP stream
- Move TCP persistence strategy files from opennms-rrd/opennms-rrd-tcponly to features/collection/persistence-tcp.
