+++ features/collection/persistence-tcp/src/main/java/org/opennms/netmgt/collection/persistence/tcp/TcpOutputStrategy.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.collection.persistence.tcponly;
+package org.opennms.netmgt.collection.persistence.tcp;
 
 import java.util.List;
