+++ features/collection/persistence-tcp/src/main/resources/META-INF/opennms/applicationContext-tcp.xml
@@ -45,24 +45,24 @@
         <constructor-arg type="java.lang.String" value="${org.opennms.rrd.usequeue}" />
     </bean>
 
-    <bean id="simpleTcpOutputStrategy" class="org.opennms.netmgt.collection.persistence.tcponly.SimpleTcpOutputStrategy" lazy-init="true">
+    <bean id="simpleTcpOutputStrategy" class="org.opennms.netmgt.collection.persistence.tcp.SimpleTcpOutputStrategy" lazy-init="true">
         <property name="host" value="${org.opennms.rrd.tcp.host}" />
         <property name="port" value="${org.opennms.rrd.tcp.port}" />
     </bean>
 
-    <bean id="queuingTcpOutputStrategy" class="org.opennms.netmgt.collection.persistence.tcponly.QueuingTcpOutputStrategy" lazy-init="true">
+    <bean id="queuingTcpOutputStrategy" class="org.opennms.netmgt.collection.persistence.tcp.QueuingTcpOutputStrategy" lazy-init="true">
         <constructor-arg ref="simpleTcpOutputStrategy" />
         <constructor-arg value="${org.opennms.rrd.queuing.queueSize}" />
     </bean>
 
-    <bean id="tcpOutputStrategyFactory" class="org.opennms.netmgt.collection.persistence.tcponly.TcpOutputStrategyFactory" />
+    <bean id="tcpOutputStrategyFactory" class="org.opennms.netmgt.collection.persistence.tcp.TcpOutputStrategyFactory" />
 
     <bean id="tcpOutputStrategy" primary="true" class="org.springframework.beans.factory.config.MethodInvokingFactoryBean">
         <property name="targetObject" ref="tcpOutputStrategyFactory"/>
         <property name="targetMethod" value="getStrategy"/>
     </bean>
 
-    <onmsgi:service ref="tcpOutputStrategy" interface="org.opennms.netmgt.collection.persistence.tcponly.TcpOutputStrategy" />
+    <onmsgi:service ref="tcpOutputStrategy" interface="org.opennms.netmgt.collection.persistence.tcp.TcpOutputStrategy" />
 
     <bean id="resourceStorageDao" class="org.opennms.netmgt.dao.support.TcpResourceStorageDao" />
 
@@ -75,7 +75,7 @@
 
     <onmsgi:service ref="rrdStrategy" interface="org.opennms.netmgt.rrd.NullRrdStrategy" />
 
-    <bean id="persisterFactory" class="org.opennms.netmgt.collection.persistence.tcponly.TcpPersisterFactory">
+    <bean id="persisterFactory" class="org.opennms.netmgt.collection.persistence.tcp.TcpPersisterFactory">
         <property name="tcpStrategy" ref="tcpOutputStrategy" />
         <property name="resourceStorageDao" ref="resourceStorageDao" />
     </bean>
