+++ opennms-rrd/pom.xml
@@ -14,7 +14,6 @@
     <module>opennms-rrd-jrobin</module>
     <module>opennms-rrd-rrdtool</module>
     <module>opennms-rrd-tcp</module>
-    <module>opennms-rrd-tcponly</module>
     <module>opennms-rrd-model</module>
   </modules>
   <build>
